+++ diive/preprocessing/outlier_detection/localsd.py
@@ -37,7 +37,7 @@ from pandas import DatetimeIndex, Series
 import diive.core.plotting.styles.LightTheme as theme
 from diive.core.base.flagbase import FlagBase
 from diive.core.plotting.plotfuncs import default_format, default_legend
-from diive.core.utils.console import detail
+from diive.core.utils.console import VERBOSE_PROGRESS, detail
 from diive.core.utils.prints import ConsoleOutputDecorator
 from diive.preprocessing.outlier_detection.common import create_daytime_nighttime_flags
 
@@ -254,7 +254,7 @@ class LocalSD(FlagBase):
         rejected = rejected[rejected].index
         n_outliers = len(rejected)
         if self.verbose:
-            detail(f"ITERATION#{iteration}{time_period}: Total found outliers: {len(rejected)} values", verbose=self.verbose)
+            detail(f"ITERATION#{iteration}{time_period}: Total found outliers: {len(rejected)} values", verbose=self.verbose, min_level=VERBOSE_PROGRESS)
         if self.showplot:
             self._plot_add_iteration(rmedian, upper_limit, lower_limit, iteration, time_period=time_period)
         return ok, rejected, n_outliers, upper_limit, lower_limit